"Add getCustomerById method to CustomerService and use it in CustomerController"
The diff includes changes to both the customer.controller.ts and customer.service.ts files. In the controller, a new method getCustomerById has been added, which takes in a response object and a parameter id of type string. This method has been decorated with the Get decorator and a route parameter of :id. In the case of an error, it returns a bad request status with an error message. If successful, it returns the customer object with a success message and an OK status.

In the service file, a new method getCustomerById has been added, which takes in an id of type string and returns a Promise of Customer type. This method uses the prisma client to find a unique customer with the given id.