"Add delete customer endpoint and service; protect delete operation with permission check"
This commit adds a new delete customer endpoint in the CustomerController that accepts id, userId, and companyId as route parameters. The corresponding service method, deleteCustomer, has also been added to the CustomerService. The deleteCustomer method first checks if the customer exists, then verifies if the user has permission to delete the customer based on their companyId. If the customer is found and the user has permission, the customer is deleted and a success message is returned. If there is an error, a failure message is returned.