"Add patch endpoint to update customer and corresponding service method; add updateCustomerDTO"
This commit adds a new endpoint to the customer controller for updating a customer with a given ID using PATCH request. It also adds a new method to the customer service for updating a customer in the database and handles the corresponding error cases. Additionally, it introduces a new DTO (Data Transfer Object) for updating customer data.