"Add getSupplierById method in supplier.service and corresponding endpoint in supplier.controller; use new 'isPublic' decorator for public endpoint"
The diff adds a new method `getSupplierById` to the `SupplierService` class, which retrieves a supplier by its unique ID. The corresponding endpoint is added to the `SupplierController` class, decorated with the new `@IsPublic()` decorator to indicate that it is a public endpoint. The `getSupplierById` method in the service uses the Prisma client to query the database and return the supplier if it exists. If the supplier is not found, an error is thrown. The endpoint returns the supplier in the response if it is found, or an error message if it is not.