"Add 'delete' permission check in SupplierService and new delete endpoint in SupplierController"
This diff adds a new parameter `method` to the `permissionUser` method in `CustomerService` and `SupplierService` to check for specific permissions. It also adds a new `deleteSupplier` method in `SupplierService` that checks for delete permission before deleting a supplier. In addition, a new `delete` endpoint is added to `SupplierController` that uses the new `deleteSupplier` method in `SupplierService`.