"Add createSupplier method to SupplierController"
This commit adds a new method, `createSupplier`, to the `SupplierController` class. The method accepts a `CreateSupplier` data transfer object (DTO) as a parameter and returns a response with the created supplier or an error message. Additionally, the `@Post` decorator is used to specify that this is a HTTP POST request.