"Add @ApiProperty decorators to customer and project DTOs"
This commit adds the `@ApiProperty` decorator from the `@nestjs/swagger` package to the properties in the `CreateCustomerDto`, `UpdateCustomerDTO`, and `ProjectOutputDto` classes. This decorator is used to provide additional metadata for the properties, which can be used by the Swagger documentation generator to generate accurate and detailed API documentation. The `IsString`, `IsNumber`, `IsBoolean`, `IsDateString`, and `IsOptional` decorators from the `class-validator` package are also used to validate the properties.