Add Product, Categories, SalesHistory, and CommentsOrNotes models to schema Add ProductService and ProductModule Update Supplier model to include products field and update relation to Company model Update schema.prisma file format for Supplier model id field
These changes include adding new models (Product, Categories, SalesHistory, and CommentsOrNotes) to the Prisma schema, as well as adding a new service (ProductService) and module (ProductModule) for handling product-related functionality. The Supplier model has also been updated to include a products field and the format of the id field in the Suplisher model has been updated in the schema file.