"Add patch endpoint to update product and corresponding service method; add updateProductDTO"
This commit adds a new endpoint to the product controller for updating a product with a given ID using PATCH request. It also adds a new method to the product service for updating a product in the database. Additionally, it introduces a new DTO (Data Transfer Object) for updating a product. The commit also includes necessary changes to the permissionUser method to check if the user has the required permissions to update the product.