Add SalesOrder model, service, and controller
The code diff includes changes to the Prisma schema to add a new model called "SalesOrder". It also includes the addition of a new service, controller, and module for SalesOrder in the NestJS application. The SalesOrder model is associated with the Company and Customer models. The binary files "dev.db" and "dev.db-journal" have also changed, likely as a result of these updates to the Prisma schema.